+++ crates/oasis7/src/bin/oasis7_chain_runtime/status_payload.rs
@@ -1,4 +1,3 @@
-use std::fs;
 use std::path::Path;
 
 use oasis7::network_tier_manifest::LoadedNetworkTierManifest;
@@ -14,12 +13,18 @@ use super::runtime_status_util::{consensus_status_to_string, now_unix_ms};
 use super::storage_metrics;
 use super::traffic_status::ChainTrafficStatus;
 use super::wasm_status::ChainWasmStatus;
+#[path = "status_payload_chain_proof.rs"]
+mod status_payload_chain_proof;
 #[path = "status_payload_network_head.rs"]
 mod status_payload_network_head;
+#[path = "status_payload_network_tier.rs"]
+mod status_payload_network_tier;
+use status_payload_chain_proof::{ChainProofStatus, build_chain_proof_status};
 pub(super) use status_payload_network_head::{
     ChainConsensusNetworkHeadStatus, ChainReadinessPolicyStatus, applied_slashing_receipt_hashes,
     build_network_head_status, pending_slashing_intent_count, readiness_policy,
 };
+use status_payload_network_tier::ChainNetworkTierStatus;
 #[path = "status_payload_observability.rs"]
 mod status_payload_observability;
 #[path = "status_payload_state_sync.rs"]
@@ -141,57 +146,6 @@ pub(super) struct ChainSyncStatus {
     pub(super) fresh_peer_count: usize,
     pub(super) stale_peer_count: usize,
     pub(super) conflicting_peer_count: usize,
-}
-
-#[derive(Debug, Serialize)]
-pub(super) struct ChainNetworkTierStatus {
-    pub(super) source_path: String,
-    pub(super) schema_version: String,
-    pub(super) tier: String,
-    pub(super) status: String,
-    pub(super) network_id: String,
-    pub(super) chain_id: String,
-    pub(super) bootstrap_peer_count: usize,
-    pub(super) governance_mode: String,
-    pub(super) validator_admission: String,
-    pub(super) target_validator_count: u64,
-    pub(super) allow_observer_nodes: bool,
-    pub(super) token_symbol: String,
-    pub(super) faucet_mode: String,
-    pub(super) reset_policy: String,
-    pub(super) value_semantics: String,
-    pub(super) rpc_ref: String,
-    pub(super) explorer_ref: String,
-    pub(super) faucet_ref: Option<String>,
-    pub(super) required_gates: Vec<String>,
-    pub(super) allowed_claims: Vec<String>,
-    pub(super) denied_claims: Vec<String>,
-}
-
-#[derive(Debug, Serialize)]
-pub(super) struct ChainProofStatus {
-    pub(super) schema_version: String,
-    pub(super) proof_contract: String,
-    pub(super) claim_boundary: String,
-    pub(super) status: String,
-    pub(super) latest_world_head_proof: Option<LatestWorldHeadProofStatus>,
-    pub(super) source_record_path: Option<String>,
-    pub(super) load_error: Option<String>,
-    pub(super) does_not_claim: Vec<String>,
-}
-
-#[derive(Debug, Serialize)]
-pub(super) struct LatestWorldHeadProofStatus {
-    pub(super) schema_version: u16,
-    pub(super) world_id: String,
-    pub(super) height: u64,
-    pub(super) execution_block_hash: String,
-    pub(super) execution_state_root: String,
-    pub(super) node_block_hash: String,
-    pub(super) action_root: String,
-    pub(super) world_head_proof_ref: String,
-    pub(super) proof_hash: String,
-    pub(super) checkpoint_ref: Option<String>,
 }
 
 #[derive(Debug, Serialize)]
@@ -430,195 +384,6 @@ pub(super) fn build_chain_p2p_status(
         deployment_mode: effective_p2p_policy.deployment_mode.as_str().to_string(),
         node_role_claim: effective_p2p_policy.node_role_claim.as_str().to_string(),
         rationale: live_p2p_recommendation.rationale.clone(),
-    }
-}
-
-fn build_chain_proof_status(execution_records_dir: Option<&Path>) -> ChainProofStatus {
-    let schema_version = "oasis7.chain_proof_status.v1".to_string();
-    let proof_contract = "WorldHeadProofV1".to_string();
-    let claim_boundary =
-        "head_execution_checkpoint_evidence_only_not_light_client_or_mainnet_readiness".to_string();
-    let does_not_claim = vec![
-        "public_testnet ready".to_string(),
-        "ready_for_live_candidate".to_string(),
-        "mainnet-grade".to_string(),
-        "light-client verified".to_string(),
-        "state proof verified".to_string(),
-        "receipt proof verified".to_string(),
-        "DA sampling verified".to_string(),
-    ];
-    let Some(execution_records_dir) = execution_records_dir else {
-        return ChainProofStatus {
-            schema_version,
-            proof_contract,
-            claim_boundary,
-            status: "unavailable".to_string(),
-            latest_world_head_proof: None,
-            source_record_path: None,
-            load_error: Some("execution_records_dir_unconfigured".to_string()),
-            does_not_claim,
-        };
-    };
-    let latest_path = execution_records_dir.join("latest.json");
-    if !latest_path.exists() {
-        return ChainProofStatus {
-            schema_version,
-            proof_contract,
-            claim_boundary,
-            status: "unavailable".to_string(),
-            latest_world_head_proof: None,
-            source_record_path: Some(latest_path.display().to_string()),
-            load_error: Some("execution_bridge_latest_record_missing".to_string()),
-            does_not_claim,
-        };
-    }
-    let bytes = match fs::read(latest_path.as_path()) {
-        Ok(bytes) => bytes,
-        Err(err) => {
-            return ChainProofStatus {
-                schema_version,
-                proof_contract,
-                claim_boundary,
-                status: "stale_or_invalid".to_string(),
-                latest_world_head_proof: None,
-                source_record_path: Some(latest_path.display().to_string()),
-                load_error: Some(format!("read latest execution record failed: {err}")),
-                does_not_claim,
-            };
-        }
-    };
-    let latest: serde_json::Value = match serde_json::from_slice(bytes.as_slice()) {
-        Ok(latest) => latest,
-        Err(err) => {
-            return ChainProofStatus {
-                schema_version,
-                proof_contract,
-                claim_boundary,
-                status: "stale_or_invalid".to_string(),
-                latest_world_head_proof: None,
-                source_record_path: Some(latest_path.display().to_string()),
-                load_error: Some(format!("parse latest execution record failed: {err}")),
-                does_not_claim,
-            };
-        }
-    };
-    let record_schema_version = latest
-        .get("schema_version")
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-    if record_schema_version < 3 {
-        return ChainProofStatus {
-            schema_version: "oasis7.chain_proof_status.v1".to_string(),
-            proof_contract,
-            claim_boundary,
-            status: "stale_or_invalid".to_string(),
-            latest_world_head_proof: None,
-            source_record_path: Some(latest_path.display().to_string()),
-            load_error: Some(format!(
-                "execution_bridge_record_schema_v{record_schema_version}_has_no_world_head_proof"
-            )),
-            does_not_claim,
-        };
-    }
-
-    let string_field = |name: &str| -> Result<String, String> {
-        latest
-            .get(name)
-            .and_then(serde_json::Value::as_str)
-            .map(str::trim)
-            .filter(|value| !value.is_empty())
-            .map(ToOwned::to_owned)
-            .ok_or_else(|| format!("{name}_missing"))
-    };
-    let height = latest
-        .get("height")
-        .and_then(serde_json::Value::as_u64)
-        .unwrap_or(0);
-    if height == 0 {
-        return ChainProofStatus {
-            schema_version: "oasis7.chain_proof_status.v1".to_string(),
-            proof_contract,
-            claim_boundary,
-            status: "stale_or_invalid".to_string(),
-            latest_world_head_proof: None,
-            source_record_path: Some(latest_path.display().to_string()),
-            load_error: Some("height_missing".to_string()),
-            does_not_claim,
-        };
-    }
-
-    let latest_world_head_proof = match (
-        string_field("world_id"),
-        string_field("execution_block_hash"),
-        string_field("execution_state_root"),
-        string_field("node_block_hash"),
-        string_field("action_root"),
-        string_field("world_head_proof_ref"),
-        string_field("world_head_proof_hash"),
-    ) {
-        (
-            Ok(world_id),
-            Ok(execution_block_hash),
-            Ok(execution_state_root),
-            Ok(node_block_hash),
-            Ok(action_root),
-            Ok(world_head_proof_ref),
-            Ok(proof_hash),
-        ) => LatestWorldHeadProofStatus {
-            schema_version: 1,
-            world_id,
-            height,
-            execution_block_hash,
-            execution_state_root,
-            node_block_hash,
-            action_root,
-            world_head_proof_ref,
-            proof_hash,
-            checkpoint_ref: latest
-                .get("checkpoint_ref")
-                .and_then(serde_json::Value::as_str)
-                .map(str::trim)
-                .filter(|value| !value.is_empty())
-                .map(ToOwned::to_owned),
-        },
-        _ => {
-            let missing_fields = [
-                "world_id",
-                "execution_block_hash",
-                "execution_state_root",
-                "node_block_hash",
-                "action_root",
-                "world_head_proof_ref",
-                "world_head_proof_hash",
-            ]
-            .into_iter()
-            .filter(|field| string_field(field).is_err())
-            .collect::<Vec<_>>();
-            return ChainProofStatus {
-                schema_version: "oasis7.chain_proof_status.v1".to_string(),
-                proof_contract,
-                claim_boundary,
-                status: "stale_or_invalid".to_string(),
-                latest_world_head_proof: None,
-                source_record_path: Some(latest_path.display().to_string()),
-                load_error: Some(format!(
-                    "latest execution record missing proof metadata: {}",
-                    missing_fields.join(",")
-                )),
-                does_not_claim,
-            };
-        }
-    };
-
-    ChainProofStatus {
-        schema_version,
-        proof_contract,
-        claim_boundary,
-        status: "available".to_string(),
-        latest_world_head_proof: Some(latest_world_head_proof),
-        source_record_path: Some(latest_path.display().to_string()),
-        load_error: None,
-        does_not_claim,
     }
 }
